RN-501: Moved SurveyGroupModel to @tupaia/database
