+++ packages/central-server/src/database/models/SurveyGroup.js
@@ -3,17 +3,9 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 
-import { DatabaseModel, DatabaseType, TYPES } from '@tupaia/database';
+import { SurveyGroupModel as CommonSurveyGroupModel } from '@tupaia/database';
 
-class SurveyGroupType extends DatabaseType {
-  static databaseType = TYPES.SURVEY_GROUP;
-}
-
-export class SurveyGroupModel extends DatabaseModel {
-  get DatabaseTypeClass() {
-    return SurveyGroupType;
-  }
-
+export class SurveyGroupModel extends CommonSurveyGroupModel {
   meditrakConfig = {
     minAppVersion: '1.6.69',
   };
