+++ packages/central-server/src/database/models/index.js
@@ -21,8 +21,8 @@ export { OptionModel as Option } from './Option';
 export { OptionSetModel as OptionSet } from './OptionSet';
 export { PermissionGroupModel as PermissionGroup } from './PermissionGroup';
 export { QuestionModel as Question } from './Question';
-export { SurveyGroupModel as SurveyGroup } from './SurveyGroup';
 export { SurveyModel as Survey } from './Survey';
+export { SurveyGroupModel as SurveyGroup } from './SurveyGroup';
 export { SurveyResponseModel as SurveyResponse } from './SurveyResponse';
 export { SurveyScreenComponentModel as SurveyScreenComponent } from './SurveyScreenComponent';
 export { SurveyScreenModel as SurveyScreen } from './SurveyScreen';
